- separation of concerns;
