+++ Proton/ruleset.xml
@@ -13,7 +13,8 @@
         <exclude name="PSR12.ControlStructures.ControlStructureSpacing.FirstExpressionLine"/> <!-- Clashes with PSR2 SpacingAfterOpenBrace -->
     </rule>
 
-    <rule ref="Proton.Commenting.InlineComment"/>
+    <rule ref="PEAR.Commenting.InlineComment"/>
+    <rule ref="Proton.Commenting.InlineCommentStartsWithSpace"/>
 
     <rule ref="Generic.Commenting.Todo"/>
     <rule ref="Generic.PHP.BacktickOperator"/>
